+++ NOTES.md
@@ -1,0 +1,8 @@
+# Notes
+
+* a session is a way to associate additional (arbitrary) data with compas data objects and/through? corresponding scene objects
+* but the scene stuff is optional
+* it is a display and/or visualization layer around the (session) data
+* every visualization/scene object references a data object
+* and overlays the data instance locally with modifiers
+* only the data and the overlay settings are required for reconstruction
